Add first functionalities to function equal
The previously declared variables var_1, var_2 and
operator are filled with the inputs of the user
and then given to the operate function which
checks with operator is used and then passes to
the correct function, either add, subtract etc.

Next step is to determine how to give the results
back to one of the higher functions and give it to
the GUI as well.